pyproject.toml
[project]
name = "crx-indicators"
version = "1.0.0"
description = "Standard indicators library for the Coinrule X execution platform"
readme = "README.md"
authors = [
    { name = "Zdenek", email = "zdenek@coinrule.com" }
]
license = "MIT"
keywords = ["trading", "crypto", "algo", "indicators"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.20.0",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "coinrule_x_indicators"}]
include = ["coinrule_x_indicators/registry.yaml"]

[tool.poetry.scripts]
coinrule_x-validate-indicators = "coinrule_x_indicators.validation:main"

[tool.poetry.urls]
Documentation = "https://github.com/coinrule_x/crx-indicators#readme"
Issues = "https://github.com/coinrule_x/crx-indicators/issues"
Source = "https://github.com/coinrule_x/crx-indicators"

[dependency-groups]
dev = [
    "pytest (<8.0.0)"
]
